\name{pedagdata}
\alias{pedagdata}
\docType{data}
\title{
  Pedagogic data set obtained during the implementation of a Sequential Agglomerative Sorting (SAS) task
}
\description{
This data set is a small artificial pedagogic data set.
}
\usage{data("pedagdata")}
\format{
An initial data frame of dimensions 10 x 30 containing the assessments on 10 stimuli provided by a panel of 10 participants during 3 steps. A cell corresponds to the set of characteristics used by a participant to describe a group of stimuli. The initial data frame is supplemented by 2 covariates providing information about the participants (Gender, Age; both categorical); and by 1 covariate providing information about the stimuli (Type; categorical).
}
\author{Margot Brard \email{margot.brard@agrocampus-ouest.fr}}

\examples{
data(pedagdata)
}

\keyword{datasets}
